refactor(test): Migrate utility functions for reuse
These functions will be used in an upcoming PR that adds additional
tests.
