+++ test/utils/index.ts
@@ -3,5 +3,6 @@ export * from "@uma/financial-templates-lib";
 
 export * from "./utils";
 export * from "./BlockchainUtils";
+export * from "./SpokePoolUtils";
 
 export { smock } from "@defi-wonderland/smock";
